Fix to the "frame time" variable - move over to using an "unsigned int" which is really just an int with the convention we will use the java unsigned comparisons when we use it (plus we rely on the usual modulo roll over for its "32 bit ness".

+++ 2COMMONPROJECTS/audio-services-test/src/test/uk/co/modularaudio/service/rackmarshalling/config/RackMarshallingTestConfig.java
@@ -82,7 +82,7 @@ public class RackMarshallingTestConfig
 	}
 
 	@Override
-	public long getCurrentUiFrameTime()
+	public int getCurrentUiFrameTime()
 	{
 		return 0;
 	}
